% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateAgeMortality}
\alias{calculateAgeMortality}
\title{calculateAgeMortality}
\usage{
calculateAgeMortality(cohortData, stage = "nonSpinup", spinupMortalityfraction)
}
\arguments{
\item{cohortData}{\code{data.table} TODO: description needed}

\item{stage}{TODO: description needed}

\item{spinupMortalityfraction}{TODO: description needed}
}
\value{
updated cohort \code{data.table}
}
\description{
TODO: description and title needed
}
