% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateCompetition}
\alias{calculateCompetition}
\title{calculateCompetition}
\usage{
calculateCompetition(cohortData, stage = "nonSpinup")
}
\arguments{
\item{cohortData}{\code{data.table} TODO: description needed}

\item{stage}{TODO: description needed}
}
\value{
updated cohort \code{data.table}
}
\description{
TODO: description and title needed
}
