% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateSumB}
\alias{calculateSumB}
\title{Calculate total biomass}
\usage{
calculateSumB(
  cohortData,
  lastReg,
  currentTime,
  successionTimestep,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData}{\code{data.table} TODO: description needed}

\item{lastReg}{TODO: description needed}

\item{currentTime}{TODO: description needed -- rename this to 'time' to match others}

\item{successionTimestep}{TODO: description needed}

\item{doAssertion}{TODO: description needed (see LandR for description)}
}
\value{
updated cohort \code{data.table}
}
\description{
Calculate the total stand biomass that does not include the new cohorts.
The new cohorts are defined as the age younger than simulation time step.
TODO: update description.
}
