% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup.R
\name{spinUp}
\alias{spinUp}
\title{spinUp}
\usage{
spinUp(
  cohortData,
  calibrate,
  successionTimestep,
  spinupMortalityfraction,
  species
)
}
\arguments{
\item{cohortData}{TODO: description needed}

\item{calibrate}{TODO: description needed}

\item{successionTimestep}{TODO: description needed}

\item{spinupMortalityfraction}{TODO: description needed}

\item{species}{TODO: description needed}

\item{fnList}{TODO: remove this once the functions are in the package!}
}
\value{
TODO: description needed
}
\description{
spinUp
}
