% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateSpeciesAttributes}
\alias{updateSpeciesAttributes}
\title{updateSpeciesAttributes}
\usage{
updateSpeciesAttributes(species, cohortData)
}
\arguments{
\item{species}{TODO: description needed}

\item{cohortData}{\code{data.table} TODO: description needed}
}
\value{
updated cohort \code{data.table}
}
\description{
TODO: description and title needed
}
