% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateSpeciesEcoregionAttributes}
\alias{updateSpeciesEcoregionAttributes}
\title{Updates the speciesEcoregion used in current year from original speciesEcoregion file}
\usage{
updateSpeciesEcoregionAttributes(speciesEcoregion, currentTime, cohortData)
}
\arguments{
\item{speciesEcoregion}{TODO: description needed}

\item{currentTime}{TODO: description needed}

\item{cohortData}{\code{data.table} TODO: description needed}
}
\value{
updated cohort \code{data.table}
}
\description{
A LANDIS-II speciesEcoregion file has a column called "year", which specifies the
year after which the speciesEcoregion values should be used. This function
compares the \code{currentTime} in the simulation with the \code{year} column in the speciesEcoregion
file and updates accordingly, if needed.
}
